Refactor Anime ID handling to use int instead of Long and add support for fetching anime by MAL ID

+++ src/main/java/com/graso/anitrack/external/anilist/dto/ResponseFetchByIdAniListDto.java
@@ -12,8 +12,8 @@ public record ResponseFetchByIdAniListDto(Data data) {
     ) {
 
         public record Media(
-                Long id,
-                Long idMal,
+                int id,
+                Integer idMal,
 
                 MediaType type,
                 Title title,
